+++ source/danger.d.ts
@@ -110,6 +110,20 @@ export interface DangerUtilsDSL {
    */
   sentence(array: Array<string>): string
 }
+/** All Text diff values will be this shape */
+export interface TextDiff {
+  /** The value before the PR's applied changes */
+  before: string
+  /** The value after the PR's applied changes */
+  after: string,
+  /** A string containing the full set of changes */
+  diff: string,
+  /** A string containing just the added lines */
+  added: string,
+  /** A string containing just the removed lines */
+  removed: string
+}
+
 /** The results of running a JSON patch */
 export interface JSONPatch {
   /** The JSON in a file at the PR merge base */
@@ -169,10 +183,8 @@ export interface GitDSL {
   /** Offers the diff for a specific file
    *
    * @param {string} filename the path to the json file
-   * @param {string[]} diffTypes An array containing: "add", "del" or "normal" - then you can only get these changes
    */
-
-  diffForFile(filename: string, diffTypes?: string[]): string | null,
+  diffForFile(filename: string): Promise<TextDiff | null>,
 
   /**
    * Provides a JSON patch (rfc6902) between the two versions of a JSON file,